requirements.txt
pillow
django-debug-toolbar
psycopg2
slugify
